requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
requests
python-dotenv
psycopg2
apscheduler